chore: connect save receip api

+++ Bookstore.Share/DTOResponses/CustomerResponse.cs
@@ -17,7 +17,7 @@ namespace Bookstore.Share.DTOResponses
         public string? LoaiKhach { get; set; }
         public string? GioiTinh { get; set; }
         public DateTime? NgaySinh { get; set; }
-        public long CongNo { get; set; }
+        public decimal CongNo { get; set; }
         public DateTime NgayTao { get; set; }
     }
 }
